sprite updates. scripting changes, additional ui
updated the poltergeist sprite to reflect some normal sprite fixes and make the background of it transparent. attempted to try and fix sorting order stuff. changed how collisions work with the raycasting shit and now i need to fix all of that but we're dealing with that later. also added the game over screen